+++ src/Codegen/ObjectLikeBuilder.php
@@ -33,7 +33,10 @@ class ObjectLikeBuilder
 
     public function __construct(string $name, string $namespace)
     {
-        $class = new ClassType($name, new PhpNamespace($namespace));
+        $class = new ClassType(
+            Escaper::escapeName($name),
+            new PhpNamespace(Escaper::escapeNamespace($namespace)) // TODO drop escape when min PHP version is 8.0+
+        );
 
         $class->addExtend(ObjectLike::class);
 
@@ -104,6 +107,8 @@ PHP
         $wrappedPhpDocType = TypeWrapper::phpDoc($type, $phpDocType);
 
         $this->class->addComment("@property {$wrappedPhpDocType} \${$name}");
+
+        $typeConverter = Escaper::escapeNamespace($typeConverter);
 
         $wrappedTypeConverter = TypeWrapper::converter($type, "new \\{$typeConverter}");
         $this->converters->addBody(/** @lang PHP */ "    '{$name}' => {$wrappedTypeConverter},");