+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.22.0
+
 ### Added
 
 - Add support for types and fields named equivalent to PHP keywords like `type Switch` or `{ print { subfield } }`.